Read client certificate file
