+++ internal/configuration/keys.go
@@ -30,3 +30,26 @@ func extractClientKey(b []byte) (key string, err error) {
 	s = strings.TrimSuffix(s, "-----END PRIVATE KEY-----")
 	return s, nil
 }
+
+func readClientCertificate(r reader) (clientCertificate string, err error) {
+	b, err := r.getFromFileOrSecretFile("OPENVPN_CLIENTCRT", constants.ClientCertificate)
+	if err != nil {
+		return "", err
+	}
+	return extractClientCertificate(b)
+}
+
+var errDecodePEMBlockClientCert = errors.New("cannot decode PEM block from client certificate")
+
+func extractClientCertificate(b []byte) (certificate string, err error) {
+	pemBlock, _ := pem.Decode(b)
+	if pemBlock == nil {
+		return "", errDecodePEMBlockClientCert
+	}
+	parsedBytes := pem.EncodeToMemory(pemBlock)
+	s := string(parsedBytes)
+	s = strings.ReplaceAll(s, "\n", "")
+	s = strings.TrimPrefix(s, "-----BEGIN CERTIFICATE-----")
+	s = strings.TrimSuffix(s, "-----END CERTIFICATE-----")
+	return s, nil
+}
