+++ internal/constants/vpnunlimited.go
@@ -9,7 +9,6 @@ import (
 //nolint:lll
 const (
 	VPNUnlimitedCertificateAuthority = "MIIEjjCCA/egAwIBAgIJAKsVbHBdakXuMA0GCSqGSIb3DQEBBQUAMIHgMQswCQYDVQQGEwJVUzELMAkGA1UECBMCTlkxETAPBgNVBAcTCE5ldyBZb3JrMR8wHQYDVQQKExZTaW1wbGV4IFNvbHV0aW9ucyBJbmMuMRYwFAYDVQQLEw1WcG4gVW5saW1pdGVkMSMwIQYDVQQDExpzZXJ2ZXIudnBudW5saW1pdGVkYXBwLmNvbTEjMCEGA1UEKRMac2VydmVyLnZwbnVubGltaXRlZGFwcC5jb20xLjAsBgkqhkiG9w0BCQEWH3N1cHBvcnRAc2ltcGxleHNvbHV0aW9uc2luYy5jb20wHhcNMTMxMjE2MTM1OTQ0WhcNMjMxMjE0MTM1OTQ0WjCB4DELMAkGA1UEBhMCVVMxCzAJBgNVBAgTAk5ZMREwDwYDVQQHEwhOZXcgWW9yazEfMB0GA1UEChMWU2ltcGxleCBTb2x1dGlvbnMgSW5jLjEWMBQGA1UECxMNVnBuIFVubGltaXRlZDEjMCEGA1UEAxMac2VydmVyLnZwbnVubGltaXRlZGFwcC5jb20xIzAhBgNVBCkTGnNlcnZlci52cG51bmxpbWl0ZWRhcHAuY29tMS4wLAYJKoZIhvcNAQkBFh9zdXBwb3J0QHNpbXBsZXhzb2x1dGlvbnNpbmMuY29tMIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDADUzS8QWGvdhLFKsEzAiq5+b0ukKjBza0k6/dmCeYTvCVqGKg/h1IAtQdVVLAkmEp0zvGH7PuOhXm7zZrCouBr/RiG4tEcoRHwc6AJmowkYERlY7+xGx3OuNrD00QceNTsan0bn78jwt0zhFNmHdoTtFjgK3oqmQYSAtbEVWYgwIDAQABo4IBTDCCAUgwHQYDVR0OBBYEFKClmYP+tMNgWagUJCCHjtaui2k+MIIBFwYDVR0jBIIBDjCCAQqAFKClmYP+tMNgWagUJCCHjtaui2k+oYHmpIHjMIHgMQswCQYDVQQGEwJVUzELMAkGA1UECBMCTlkxETAPBgNVBAcTCE5ldyBZb3JrMR8wHQYDVQQKExZTaW1wbGV4IFNvbHV0aW9ucyBJbmMuMRYwFAYDVQQLEw1WcG4gVW5saW1pdGVkMSMwIQYDVQQDExpzZXJ2ZXIudnBudW5saW1pdGVkYXBwLmNvbTEjMCEGA1UEKRMac2VydmVyLnZwbnVubGltaXRlZGFwcC5jb20xLjAsBgkqhkiG9w0BCQEWH3N1cHBvcnRAc2ltcGxleHNvbHV0aW9uc2luYy5jb22CCQCrFWxwXWpF7jAMBgNVHRMEBTADAQH/MA0GCSqGSIb3DQEBBQUAA4GBALkWhfw7SSV7it+ZYZmT+cQbExjlYgQ40zae2J2CqIYACRcfsDHvh7Q+fiwSocevv2NE0dWi6WB2H6SiudYjvDvubAX/QbzfBxtbxCCoAIlfPCm8xOnWFN7TUJAzWwHJkKgEnu29GZHu2x8J+7VeDbKH5RTMHHe8FkSxh6Zz/BMN"
-	VPNUnlimitedCertificate          = "MIIFXTCCBMagAwIBAgIBATANBgkqhkiG9w0BAQUFADCB4DELMAkGA1UEBhMCVVMxCzAJBgNVBAgTAk5ZMREwDwYDVQQHEwhOZXcgWW9yazEfMB0GA1UEChMWU2ltcGxleCBTb2x1dGlvbnMgSW5jLjEWMBQGA1UECxMNVnBuIFVubGltaXRlZDEjMCEGA1UEAxMac2VydmVyLnZwbnVubGltaXRlZGFwcC5jb20xIzAhBgNVBCkTGnNlcnZlci52cG51bmxpbWl0ZWRhcHAuY29tMS4wLAYJKoZIhvcNAQkBFh9zdXBwb3J0QHNpbXBsZXhzb2x1dGlvbnNpbmMuY29tMB4XDTIxMDYwMTE3NDU1M1oXDTMxMDUzMDE3NTA1M1owgeUxCzAJBgNVBAYTAlVTMQswCQYDVQQIEwJOWTERMA8GA1UEBxMITmV3IFlvcmsxHzAdBgNVBAoTFlNpbXBsZXggU29sdXRpb25zIEluYy4xIzAhBgNVBAsTGml0LW1pbC52cG51bmxpbWl0ZWRhcHAuY29tMS0wKwYDVQQDEyRLUzItN2M4YTg5MzNiN2VmOTZiOTQ4Yjg3YjZhN2I2MDNlZTgxLTArBgNVBCkTJEVCMjI5MjY3LUExRTItNDg3Ni1BNTE5LTZGNzZEQURFNDRGNTESMBAGCSqGSIb3DQEJARYDWFhYMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAzzUL+7RnimM6gXHILVNlZypg0/HRNAAkMSmEvrQ+/EI4gyAuDJLEIDT9Pr+wBQv7Bkh8xyJikZ6oClbjEfY+JImUvjuJaeeaSWJfPzBLDtW4Kb0N4l4I5OlmZrmgMYT/8TnXZ8c7T7q9MSwR3cwVetlEc8KZh5DNEcioyEdP7/jQ5tOTcpQdopMu+PyDBeiuKiQbS+O0kyIw2yaBxMIEZaXvyr5DuPPjHa4zcUAFPWfLt4ea2/HhOUE46X/oMUqzErM0nkqYFKyCYdhBFDtr5b3BF58BQ79Lyfd2hUsqqHNHaPCx1w9spE6UBdUbpVCFg62w1xd/DgJ3LrjutOSOxwIDAQABo4IBmjCCAZYwCQYDVR0TBAIwADAtBglghkgBhvhCAQ0EIBYeRWFzeS1SU0EgR2VuZXJhdGVkIENlcnRpZmljYXRlMB0GA1UdDgQWBBSCiJtbYMNuHtRV6XbBYqf37sbPlDCCARcGA1UdIwSCAQ4wggEKgBSgpZmD/rTDYFmoFCQgh47WrotpPqGB5qSB4zCB4DELMAkGA1UEBhMCVVMxCzAJBgNVBAgTAk5ZMREwDwYDVQQHEwhOZXcgWW9yazEfMB0GA1UEChMWU2ltcGxleCBTb2x1dGlvbnMgSW5jLjEWMBQGA1UECxMNVnBuIFVubGltaXRlZDEjMCEGA1UEAxMac2VydmVyLnZwbnVubGltaXRlZGFwcC5jb20xIzAhBgNVBCkTGnNlcnZlci52cG51bmxpbWl0ZWRhcHAuY29tMS4wLAYJKoZIhvcNAQkBFh9zdXBwb3J0QHNpbXBsZXhzb2x1dGlvbnNpbmMuY29tggkAqxVscF1qRe4wEwYDVR0lBAwwCgYIKwYBBQUHAwIwCwYDVR0PBAQDAgeAMA0GCSqGSIb3DQEBBQUAA4GBAFpRuj2vyLjDkUGnS6vmfCNZapEJHjLs8yDnqXXNPmyEyOsn1ivvW9uoWqOlA0VThkwoMb8+jAZCSwHavfoULBRPAQzddy5WCuG1OkhCs2eK/JUfIKEx8/XSf7hFJoEbBwRA4U7napmtaWGeVnV4SssPVIFGWuW8KM5Jv1hC9jys"
 )
 
 func VPNUnlimitedCountryChoices() (choices []string) {
